Downgraded JUnit from 5 to 4
The test for the HtmlHelper must mock the Html.escapeHtml method, which
is static. To do this, it needs PowerMock.

PowerMock does not support JUnit 5 since the API for JUnit extensions
changed between JUnit 4 and JUnit 5.

Therefore, to have all the nice features of PowerMock, JUnit has to stay
at version 4.